[Feat] Lobby2Player, Lobby2OVRCameraRig 작성
- WesternLobby에서 사용할 플레이어 프리팹을 작성
- BeerInteraction, OakBarrelInteraction, FirstMoveAttackPlayer, PlayerGoldRushInteraction, PlayerTumbleweedInteraction 부착
- 더불어 KeyBoardManager 한글 입력 부분 제외 주석 추가
