+++ Assets/Resources/Lobby2OVRCameraRig.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ea955061a965c6742a33f1f02295947e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
