+++ Assets/Resources/Lobby2Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7e38fdce383b2574485453a9d7abd347
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
