+++ Assets/_Scripts/Lobby2_Interact/FindingGold/PlayerTumbleweedInteraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fbc8deb4a9894544386711a5edc91a1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
