--- Assets/_Scripts/Player/PlayerGoldRushInteraction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2723cb4d80230e4198e528796b2cc9d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
